requirements_tests.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
python-decouple==3.8
python-dotenv==1.1.1
qrcode==8.2
reportlab==4.4.3
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
